Add files via upload
Objective: To create a distributed file system for reliable and secure file storage. 1. Added support for GET, PUT and LIST commands. (PUT command divides each file into 4 pieces and stores on 4 different servers, so that it can be retrieved even if one server is not active. LIST command inquires what files are present on each of the 4 servers and also determines if a file can be reconstructed using the information obtained as part of LIST response. GET command downloads all available pieces of a file from all available DFS.)